Create README.md
The tool is built using advanced python modules that support kernel and terminal level of access and operation rights. The tool provides a wide range of options to check vulnerabilities and perform penetration testing on the selected system or networks. The options are categorised into network related tests and system related tests. The tool is capable of performing penetration tests from a low security system to a highly secured firewall system.